+++ .internal/assemblyscript/@gl/api/w2h/host.ts
@@ -4,6 +4,7 @@ import * as debug from "./debug";
 import * as filters from "./filters";
 import * as lights from "./lights";
 import * as map from "./map";
+import * as markers from "./markers";
 import * as npc from "./npc";
 import * as particles from "./particles";
 import * as physics from "./physics";
@@ -26,6 +27,7 @@ export {
   filters,
   lights,
   map,
+  markers,
   npc,
   particles,
   physics,